+++ src/Errors.ts
@@ -1,43 +1,47 @@
 import type { HttpResponse, RokuMessages } from './RokuDeploy';
 import type * as requestType from 'request';
 
+export interface RequestResult {
+    response: requestType.Response;
+    body: any;
+}
 export class InvalidDeviceResponseCodeError extends Error {
-    constructor(message: string, public results?: any) {
+    constructor(message: string, public results?: RequestResult) {
         super(message);
         Object.setPrototypeOf(this, InvalidDeviceResponseCodeError.prototype);
     }
 }
 
 export class UnauthorizedDeviceResponseError extends Error {
-    constructor(message: string, public results?: any) {
+    constructor(message: string, results?: any) {
         super(message);
         Object.setPrototypeOf(this, UnauthorizedDeviceResponseError.prototype);
     }
 }
 
 export class UnparsableDeviceResponseError extends Error {
-    constructor(message: string, public results?: any) {
+    constructor(message: string, results?: any) {
         super(message);
         Object.setPrototypeOf(this, UnparsableDeviceResponseError.prototype);
     }
 }
 
 export class FailedDeviceResponseError extends Error {
-    constructor(message: string, public results?: any) {
+    constructor(message: string, results?: any) {
         super(message);
         Object.setPrototypeOf(this, FailedDeviceResponseError.prototype);
     }
 }
 
 export class UnknownDeviceResponseError extends Error {
-    constructor(message: string, public results?: any) {
+    constructor(message: string, results?: any) {
         super(message);
         Object.setPrototypeOf(this, UnknownDeviceResponseError.prototype);
     }
 }
 
 export class CompileError extends Error {
-    constructor(message: string, public results: any, public rokuMessages: RokuMessages) {
+    constructor(message: string, results: any, rokuMessages: RokuMessages) {
         super(message);
         Object.setPrototypeOf(this, CompileError.prototype);
     }