+++ src/RokuDeployOptions.ts
@@ -1,10 +1,10 @@
-import type { LogLevel } from './Logger';
+import type { LogLevel } from '@rokucommunity/logger';
 
 export interface RokuDeployOptions {
     /**
-     * Path to a bsconfig.json project file
+     * The working directory where the command should be executed
      */
-    project?: string;
+    cwd?: string;
 
     /**
      * A full path to the folder where the zip/pkg package should be placed
@@ -46,29 +46,10 @@ export interface RokuDeployOptions {
     files?: FileEntry[];
 
     /**
-     * Set this to true to prevent the staging folder from being deleted after creating the package
-     * @default false
-     * @deprecated use `retainStagingDir` instead
-     */
-    retainStagingFolder?: boolean;
-
-    /**
-     * Set this to true to prevent the staging folder from being deleted after creating the package
-     * @default false
-     */
-    retainStagingDir?: boolean;
-
-    /**
      * Should the zipped package be retained after deploying to a roku. If false, this will delete the zip after a deployment.
      * @default true
      */
     retainDeploymentArchive?: boolean;
-
-    /**
-     * The path where roku-deploy should stage all of the files right before being zipped. defaults to ${outDir}/.roku-deploy-staging
-     * @deprecated since 3.9.0. use `stagingDir` instead
-     */
-    stagingFolderPath?: string;
 
     /**
      * The path where roku-deploy should stage all of the files right before being zipped. defaults to ${outDir}/.roku-deploy-staging
@@ -108,6 +89,11 @@ export interface RokuDeployOptions {
     remotePort?: number;
 
     /**
+     * The directory where screenshots should be saved. Will use the OS temp directory by default
+     */
+    screenshotDir?: string;
+
+    /**
      * The request timeout duration in milliseconds. Defaults to 150000ms (2 minutes 30 seconds).
      * This is mainly useful for preventing hang ups if the Roku loses power or restarts due to a firmware bug.
      * This is applied per network request to the device and does not apply to the total time it takes to completely execute a call to roku-deploy.
@@ -142,16 +128,6 @@ export interface RokuDeployOptions {
     devId?: string;
 
     /**
-     * If true we increment the build number to be a timestamp in the format yymmddHHMM
-     */
-    incrementBuildNumber?: boolean;
-
-    /**
-     * If true we convert to squashfs before creating the pkg file
-     */
-    convertToSquashfs?: boolean;
-
-    /**
      * If true, the publish will fail on compile error
      */
     failOnCompileError?: boolean;
@@ -165,7 +141,7 @@ export interface RokuDeployOptions {
     /**
      * If true, the previously installed dev channel will be deleted before installing the new one
      */
-    deleteInstalledChannel?: boolean;
+    deleteDevChannel?: boolean;
 
     /**
      * Overrides for values used during the zip upload process. You probably don't need to change these...
@@ -184,4 +160,4 @@ export interface RokuDeployOptions {
     };
 }
 
-export type FileEntry = (string | { src: string | string[]; dest?: string });
+export type FileEntry = (string | { src: string[] | string; dest?: string });